refactor: 함수, 상수 분리
